e/mail-relay/requirements_all.txt
authres==1.0.1
chardet==2.3.0
Django==1.8
django-classy-tags==0.7.2
django-debug-toolbar==1.4
django-excel-response==1.0
django-filter==0.11.0
django-mptt==0.8.3
django-nyt==0.9.9
django-redis==3.6.2
django-sekizai==0.9.0
django-simple-captcha==0.5.1
django-states==1.6.5
djangorestframework==3.1.3
dnspython==1.15.0
ecdsa==0.13
Fabric==1.10.1
Flask==0.10.1
gevent==1.0.2
greenlet==0.4.6
ipaddr==2.1.11
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
jsonfield==2.0.2
Markdown==2.6.6
MarkupSafe==0.23
meld3==1.0.2
paramiko==1.16.0
phpserialize==1.3
#Pillow==3.1.0
psycopg2==2.6
pyClamd==0.3.17
pycrypto==2.6.1
pydns==2.3.6
pyspf==2.0.11
python-dateutil==2.4.2
redis==2.10.5
requests==2.10.0
six==1.10.0
sorl-thumbnail==12.3
SPF2IP==1.0.4
sqlparse==0.1.18
supervisor==3.1.3
urllib3==1.10.4
Werkzeug==0.11.3
wiki==0.1
